+++ Stats/Modules/CPU/CPU.swift
@@ -16,8 +16,6 @@ class CPU: Module {
     var submenu: NSMenu = NSMenu()
     var active: Observable<Bool>
     var available: Observable<Bool>
-    var color: Observable<Bool>
-    var label: Observable<Bool>
     var hyperthreading: Observable<Bool>
     var reader: Reader = CPUReader()
     
@@ -29,15 +27,15 @@ class CPU: Module {
         self.active = Observable(defaults.object(forKey: name) != nil ? defaults.bool(forKey: name) : true)
         self.hyperthreading = Observable(defaults.object(forKey: "\(name)_hyperthreading") != nil ? defaults.bool(forKey: "\(name)_hyperthreading") : true)
         self.widgetType = defaults.object(forKey: "\(name)_widget") != nil ? defaults.float(forKey: "\(name)_widget") : Widgets.Mini
-        self.color = Observable(defaults.object(forKey: "\(name)_color") != nil ? defaults.bool(forKey: "\(name)_color") : false)
-        self.label = Observable(defaults.object(forKey: "\(name)_label") != nil ? defaults.bool(forKey: "\(name)_label") : true)
-        initMenu()
-        initWidget()
         
         if self.widgetType == Widgets.BarChart {
             (self.reader as! CPUReader).perCoreMode = true
             (self.reader as! CPUReader).hyperthreading = self.hyperthreading.value
+            self.reader.read()
         }
+        
+        initWidget()
+        initMenu()
     }
     
     func initMenu() {
@@ -71,14 +69,6 @@ class CPU: Module {
         hyperthreading.state = self.hyperthreading.value ? NSControl.StateValue.on : NSControl.StateValue.off
         hyperthreading.target = self
         
-        let color = NSMenuItem(title: "Color", action: #selector(toggleColor), keyEquivalent: "")
-        color.state = self.color.value ? NSControl.StateValue.on : NSControl.StateValue.off
-        color.target = self
-        
-        let label = NSMenuItem(title: "Label", action: #selector(toggleLabel), keyEquivalent: "")
-        label.state = self.label.value ? NSControl.StateValue.on : NSControl.StateValue.off
-        label.target = self
-        
         submenu.addItem(mini)
         submenu.addItem(chart)
         submenu.addItem(chartWithValue)
@@ -86,14 +76,14 @@ class CPU: Module {
         
         submenu.addItem(NSMenuItem.separator())
         
+        if let view = self.view as? Widget {
+            for widgetMenu in view.menus {
+                submenu.addItem(widgetMenu)
+            }
+        }
+        
         if self.widgetType == Widgets.BarChart {
             submenu.addItem(hyperthreading)
-        }
-        if self.widgetType == Widgets.BarChart || self.widgetType == Widgets.ChartWithValue || self.widgetType == Widgets.Chart {
-            submenu.addItem(label)
-        }
-        if self.widgetType == Widgets.Mini || self.widgetType == Widgets.ChartWithValue {
-            submenu.addItem(color)
         }
         
         menu.submenu = submenu
@@ -160,19 +150,5 @@ class CPU: Module {
         self.defaults.set(sender.state == NSControl.StateValue.on, forKey: "\(name)_hyperthreading")
         self.hyperthreading << (sender.state == NSControl.StateValue.on)
         (self.reader as! CPUReader).hyperthreading = sender.state == NSControl.StateValue.on
-    }
-    
-    @objc func toggleColor(_ sender: NSMenuItem) {
-        sender.state = sender.state == NSControl.StateValue.on ? NSControl.StateValue.off : NSControl.StateValue.on
-        self.defaults.set(sender.state == NSControl.StateValue.on, forKey: "\(name)_color")
-        self.color << (sender.state == NSControl.StateValue.on)
-    }
-    
-    @objc func toggleLabel(_ sender: NSMenuItem) {
-        sender.state = sender.state == NSControl.StateValue.on ? NSControl.StateValue.off : NSControl.StateValue.on
-        self.defaults.set(sender.state == NSControl.StateValue.on, forKey: "\(name)_label")
-        self.active << false
-        self.label << (sender.state == NSControl.StateValue.on)
-        self.active << true
     }
 }